requirements.txt
aws-shell==0.2.0
awscli==1.14.17
boto3==1.5.6
botocore==1.8.20
colorama==0.3.7
configobj==5.0.6
dj-database-url==0.4.2
Django==1.11.5
django-braces==1.11.0
django-ckeditor==5.3.0
django-contrib-comments==1.8.0
django-cuser==2017.3.16
django-js-asset==0.1.1
django-model-utils==3.0.0
django-modeltranslation==0.12.1
django-mptt==0.9.0
django-recaptcha==1.3.1
django-star-ratings==0.6.0
django-storages==1.6.5
djangorestframework==3.7.0
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
olefile==0.44
Pillow==4.2.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
pyasn1==0.4.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.12
six==1.11.0
swapper==1.1.0
wcwidth==0.1.7
